pyproject.toml
[tool.poetry]
name = "nautobot-route-tracking"
version = "1.0.0"
description = "Nautobot plugin for tracking routing table entries from network devices"
authors = ["Thomas <thomas@networktocode.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/tcheval/nautobot-route-tracking"
repository = "https://github.com/tcheval/nautobot-route-tracking"
keywords = ["nautobot", "nautobot-plugin", "routing", "route-tracking", "napalm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Plugins",
    "Framework :: Django",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
]
packages = [{ include = "nautobot_route_tracking" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
nautobot = "^3.0.6"
nornir = "^3.4.0"
nornir-nautobot = "^4.0.0"
nautobot-plugin-nornir = "^3.0.0"
nornir-napalm = "^0.5.0"
napalm = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
coverage = "^7.4.0"
ruff = "^0.2.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "D",     # pydocstring
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "DJ",    # flake8-django
    "SIM",   # flake8-simplify
    "RUF",   # Ruff-specific rules
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D106",   # Missing docstring in public nested class
    "D203",   # 1 blank line required before class docstring
    "D213",   # Multi-line docstring summary should start at the second line
    "D401",   # First line of docstring should be in imperative mood
    "D406",   # Section name should end with a newline
    "D407",   # Missing dashed underline after section
    "D413",   # Missing blank line after last section
    "RUF012", # Mutable class attributes (Django admin/Meta classes are standard patterns)
    "SIM102", # Nested if statements (sometimes more readable)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "S101"]
"**/migrations/*" = ["D", "E501", "I"]

[tool.ruff.lint.isort]
known-first-party = ["nautobot_route_tracking"]
known-third-party = ["nautobot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
DJANGO_SETTINGS_MODULE = "tests.test_settings"

[tool.coverage.run]
source = ["nautobot_route_tracking"]
branch = true
omit = [
    "*/migrations/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 80
